requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.55.3
numpy
pydantic==2.5.1
python-dotenv==1.0.0
faiss-cpu
langchain
langchain-core
python-docx
dotenv-python